requirements.txt
Django==2.0
djangorestframework==3.7.7
pycrypto==2.6.1
uwsgi
request